% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residrill_import.R
\name{drillingresistance_plot}
\alias{drillingresistance_plot}
\title{drillingresistance_plot}
\usage{
drillingresistance_plot(df)
}
\arguments{
\item{df}{data.frame of residrill data}
}
\description{
drillingresistance_plot
}

