% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residrill_import.R
\name{residrill_import}
\alias{residrill_import}
\title{residrill_import}
\usage{
residrill_import(filename)
}
\arguments{
\item{filename}{}
}
\description{
Functions for reading in residrill output
}

